Split ViewHierarchyNodeTests to separate file
